[serve] move test from test_grpc to test_proxy (#53933)
## Why are these changes needed?

Move serve test to a different file.
